Modiify modules & fix linter errors
