+++ modules/book.js
@@ -1,6 +1,6 @@
 export default class Book {
-    constructor(title, author) {
-        this.title = title;
-        this.author = author;
-    }
-};
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
