--- modules/clear.js
@@ -1,6 +1,0 @@
-const refresh = (title, author) => {
-  title.value = '';
-  author.value = '';
-};
-
-export { refresh };
